+++ lib/lightning_web/live/workflow_live.ex
@@ -199,6 +199,13 @@ defmodule LightningWeb.WorkflowLive do
   end
 
   @impl true
+  def handle_event("copied-to-clipboard", _, socket) do
+    {:noreply,
+     socket
+     |> put_flash(:info, "Copied webhook URL to clipboard")}
+  end
+
+  @impl true
   def handle_event("create-workflow", _, socket) do
     {:ok, %Workflows.Workflow{id: workflow_id}} =
       Workflows.create_workflow(%{project_id: socket.assigns.project.id})